fix: sync selected topic header
