+++ e2e/index.test.ts
@@ -1,9 +1,6 @@
-import { Page, expect } from "@playwright/test";
-
-import { basePath } from "~e2e/helper/constants";
-import { wait } from "~shared/time";
-
 import { test } from "./helper/base-test";
+import { createAndGotoRoom, createAndGotoTopic } from "./helper/common-actions";
+import { basePath } from "./helper/constants";
 
 test("find login button", async ({ page }) => {
   await page.goto(basePath);
@@ -16,64 +13,16 @@ test("view spaces", async ({ page, db, auth }) => {
   await page.waitForSelector(`text=${db.space.name}`);
 });
 
-type RoomProps = { spaceName: string; roomName: string };
-async function createAndNavigateToRoom(page: Page, { spaceName, roomName }: RoomProps) {
-  await page.goto(basePath);
-  await page.click("text='New Room'");
-  await page.click("text=Room name", { force: true });
-  await page.keyboard.type(roomName);
-  await page.click("text='Space'", { force: true });
-  await page.click(`[role="option"]:has-text("${spaceName}")`);
-  await page.click("text='Create'");
-
-  await page.waitForNavigation({ url: (url) => url.pathname.startsWith("/space") });
-  await page.waitForSelector(`text='${roomName}'`);
-}
-
 test("create a room", async ({ page, db, auth }) => {
   await auth.login(db.user2);
-  await createAndNavigateToRoom(page, { spaceName: db.space.name, roomName: "Roomba" });
-});
-
-async function createTopic(page: Page, { topicName, ...roomProps }: { topicName: string } & RoomProps) {
-  await createAndNavigateToRoom(page, roomProps);
-  await page.click("text=New Topic");
-  await page.waitForFunction(
-    () => document.activeElement instanceof HTMLElement && document.activeElement.contentEditable
-  );
-
-  // playwright can be faster in picking up focus than our app in handling keystrokes, so we wait
-  await wait(1000);
-  await page.keyboard.type(topicName);
-  await page.keyboard.press("Enter");
-
-  await page.waitForSelector(`h3:text("${topicName}")`);
-  expect((await page.$$(`text='${topicName}'`)).length).toBe(2);
-  // should only find the new topic button, not a topic with that name, as we just renamed it
-  expect((await page.$$("text=New Topic")).length).toBe(1);
-}
-
-test("create a topic", async ({ page, db, auth }) => {
-  await auth.login(db.user2);
-  await createTopic(page, { spaceName: db.space.name, roomName: "Moor", topicName: "hot takes" });
-});
-
-test("delete a topic", async ({ page, db, auth }) => {
-  await auth.login(db.user2);
-  const topicName = "not for long";
-  await createTopic(page, { spaceName: db.space.name, roomName: "DoomRoom", topicName });
-
-  await page.click("button:right-of(:text('Close Topic'))");
-  await page.click("text=Delete topic");
-  await page.click("text='Delete'");
-  await page.waitForSelector(`text='${topicName}'`, { state: "hidden" });
+  await createAndGotoRoom(page, { spaceName: db.space.name, roomName: "Roomba" });
 });
 
 const COMPOSER_SELECTOR = '[class^="RichEditor__UIEditorContent"]';
 
 test("send a message", async ({ page, db, auth }) => {
   await auth.login(db.user2);
-  await createTopic(page, { spaceName: db.space.name, roomName: "Mailbox", topicName: "hot takes" });
+  await createAndGotoTopic(page, { spaceName: db.space.name, roomName: "Mailbox", topicName: "hot takes" });
   await page.click(COMPOSER_SELECTOR);
   const message = "What is happening";
   await page.keyboard.type(message);
@@ -83,7 +32,7 @@ test("send a message", async ({ page, db, auth }) => {
 
 test("replying to a request-response marks it as done", async ({ page, db, auth }) => {
   await auth.login(db.user2);
-  await createTopic(page, { spaceName: db.space.name, roomName: "ResponsiveRoom", topicName: "cold takes" });
+  await createAndGotoTopic(page, { spaceName: db.space.name, roomName: "ResponsiveRoom", topicName: "cold takes" });
 
   await page.click(COMPOSER_SELECTOR);
   await page.keyboard.type("What is happening @u");
@@ -103,7 +52,7 @@ test("replying to a request-response marks it as done", async ({ page, db, auth 
 
 test("sending a message with tasks for read and response, asks for the latter", async ({ page, db, auth }) => {
   await auth.login(db.user2);
-  await createTopic(page, { spaceName: db.space.name, roomName: "Mailbox", topicName: "hot takes" });
+  await createAndGotoTopic(page, { spaceName: db.space.name, roomName: "Mailbox", topicName: "hot takes" });
   await page.click(COMPOSER_SELECTOR);
 
   await page.keyboard.type("What is happening @u");
